refactor: move ripple to a hook
